The task section broken updated with  fixes.
The task section was broken, when you clicked to add the item it would immediately try and find the task directory instead of doing index.php/tasks like it is supposed to. IE locahost/index.php/tasks and I click to add a task it would go to /task instead. Since I could not find a blade with task name. I redid the routes and the form post items to do 'tasks' instead to ensure we got to the proper blade. This fixed the issue I was having when I did these changes.